pyproject.toml
[tool.poetry]
name = "fastapi-prj"
version = "0.1.0"
description = ""
authors = ["arthur <myemail@email.org>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
fastapi = {extras = ["all"], version = "^0.103.1"}
pyright = "^1.1.328"
python-language-server = "^0.36.2"
sqlalchemy = "^2.0.21"
asyncpg = "^0.28.0"
alembic = "^1.12.0"
pydantic-settings = "^2.0.3"
passlib = "^1.7.4"
python-jose = "^3.3.0"
fastapi-cache2 = {extras = ["redis"], version = "^0.2.1"}
celery = "^5.3.4"
flower = "^2.0.1"
pillow = "^10.0.1"
sqladmin = {extras = ["full"], version = "^0.15.1"}
pytest = "^7.4.2"
pytest-asyncio = "^0.21.1"
pudb = "^2020.1"
python-json-logger = "^2.0.7"
gunicorn = "^21.2.0"
prometheus-fastapi-instrumentator = "^6.1.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
